backend/requirements.txt
# Core requirements
Flask>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
boto3>=1.18.0

# Embedding and vector storage
cohere>=4.0.0
chromadb>=0.4.0
redis>=4.0.0

# Document processing
PyPDF2>=2.0.0
python-docx>=0.8.11
beautifulsoup4>=4.10.0

# Utilities
tqdm>=4.62.0
pandas>=1.3.0
numpy>=1.20.0

# Optional NLP enhancements
spacy>=3.0.0
nltk>=3.6.0 